+++ CHANGELOG.md
@@ -1,9 +1,14 @@
 # Release notes
 
+## Next release
+
 - Consolidated track options of horizontal and vertical tracks
 - Set background of new heatmap tracks in a combined track to `transparent` when added via the track context menu.
 - Fixed bug where heatmap labels could not be hidden.
 - The position of labels and colorbars for split heatmaps can now be changed.
+- Use itemRgb field in bed files as the default coloring for bedlike tracks
+
+_[Detailed changes since v1.9.2](https://github.com/higlass/higlass/compare/v1.9.2...develop)_
 
 ## v1.9.2